Merge 531 upstream Monado commits into dev, resolving 44 conflicted
files across all subsystems. Key upstream changes integrated:

- Compositor CMakeLists split into subdirectories (main/, multi/, etc.)
- IPC proto.json split into per-domain files (proto/*.json)
- oxr_extension_support.h replaced by Python generator
- Action/binding system refactored with reference counting
- Vulkan helper restructuring with generated .inc files
- New drivers: PSVR2, SolarXR, Rift Sensor
- xrt_future API, space queries, callback system
- u_string_list replaced by immutable u_extension_list + builder
- vk_bundle_queue abstraction for present queues
- Hand tracking extracted to oxr_hand_tracking.h/c

All dev innovations preserved:
- D3D11 native compositor and service compositor
- Display processor vtable abstraction
- Leia/sim_display drivers
- Kooima FOV projection
- XR_EXT_win32_window_binding extension
- macOS/MoltenVK portability fixes
- SR view poses and display mode IPC extensions

+++ src/xrt/auxiliary/util/u_distortion_mesh.c
@@ -381,10 +381,10 @@ u_compute_distortion_ns_meshgrid(
 
 	struct xrt_fov fov = values->fov[view];
 
-	float left_ray_bound = tan(fov.angle_left);
-	float right_ray_bound = tan(fov.angle_right);
-	float up_ray_bound = tan(fov.angle_up);
-	float down_ray_bound = tan(fov.angle_down);
+	float left_ray_bound = tanf(fov.angle_left);
+	float right_ray_bound = tanf(fov.angle_right);
+	float up_ray_bound = tanf(fov.angle_up);
+	float down_ray_bound = tanf(fov.angle_down);
 
 	float u_eye = math_map_ranges(bearing.x, left_ray_bound, right_ray_bound, 0, 1);
 	float v_eye = math_map_ranges(bearing.y, down_ray_bound, up_ray_bound, 0, 1);
